remotecontrol/server/controlcenter/requirements.txt
click==6.7
eventlet==0.22.1
Flask==0.12.2
Flask-SocketIO==2.9.6
Flask-OAuthlib==0.9.4
greenlet==0.4.13
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-engineio==2.0.4
python-socketio==1.9.0
six==1.11.0
Werkzeug==0.14.1
PyJWT==1.6.4
